clean up kibana link generator
